python/requirements.txt
google-generativeai
google-genai
python-dotenv
requests